requirements.txt
requests
streamlit
pymupdf
python-dotenv
bcrypt
pandas
faiss-cpu
sentence-transformers